The resource bug is STILL HERE!
