--- src/main/java/com/xbuilders/engine/Difficulty.java
@@ -1,5 +1,0 @@
-package com.xbuilders.engine;
-
-public enum Difficulty {
-    EASY, NORMAL, HARD
-}
