+++ src/main/java/com/xbuilders/engine/server/Difficulty.java
@@ -1,0 +1,5 @@
+package com.xbuilders.engine.server;
+
+public enum Difficulty {
+    EASY, NORMAL, HARD
+}
