+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.xbuilders.engine.ResourceList
+Main-Class: com.xbuilders.Main
